Move snakeCaseify logic into new caseConversion module
