+++ packages/c2pa-utilities/src/index.ts
@@ -9,3 +9,4 @@
 
 export * from './settings.js';
 export * from './fetchWithRetry.js';
+export * from './caseConversion.js';
